:construction: wip: add select field logic
