+++ views/pools/find-pool-modal/find-pool-modal.types.ts
@@ -8,3 +8,7 @@ export interface FindPoolModalProps {
 export interface FindPoolForm {
   tokens: ReadonlyArray<CoinData>;
 }
+
+export interface SelectTokenProps {
+  index: number;
+}
