--- views/pools/find-pool-modal/select-coins/input/input.types.ts
@@ -1,8 +1,0 @@
-export interface InputProps {
-  index: number;
-}
-
-export interface HeaderInfoProps {
-  type: string;
-  decimals: number;
-}
